+++ global.d.ts
@@ -1,5 +1,10 @@
 import { PrismaClient } from "@prisma/client";
 
+type Messages =
+| typeof import("./public/locales/eng/common.json")
+| typeof import("./public/locales/ru/common.json")
+| typeof import("./public/locales/pl/common.json");
+
 declare global {
   namespace NodeJS {
     interface Global {
@@ -9,6 +14,7 @@ declare global {
   interface GtagWindow extends Window {
     gtag: any;
   }
+  interface IntlMessages extends Messages {}
 }
 
 declare var window: GtagWindow & typeof globalThis;