+++ src/app/[locale]/layout.tsx
@@ -1,10 +1,11 @@
+import { NextIntlClientProvider, useMessages } from "next-intl";
 import { Footer } from "@/components/Footer";
 import { Header } from "@/components/Header";
 import { NotoSans } from "@/fontLoader.ts";
 import { css } from "~/styled-system/css";
+import { Box } from "~/styled-system/jsx";
 
 import "../globals.css";
-import { Box } from "~/styled-system/jsx";
 
 export const metadata = {
   title: "Next.js",
@@ -20,17 +21,23 @@ export default function RootLayout({
     locale: string;
   };
 }) {
+  const messages = useMessages();
+
   return (
     <html lang={locale} className={`${NotoSans.variable}`}>
-      <body className={css({ fontFamily: "noto" })}>
-        <Box minHeight="100vh">
+      <NextIntlClientProvider locale={locale} messages={messages}>
+        <body className={css({ fontFamily: "noto" })}>
           <Header />
-          <main className={css({ height: 0, minHeight: "calc(100vh - 85px)" })}>
-            {children}
-          </main>
-        </Box>
-        <Footer />
-      </body>
+          <Box minHeight="100vh">
+            <main
+              className={css({ height: 0, minHeight: "calc(100vh - 85px)" })}
+            >
+              {children}
+            </main>
+          </Box>
+          <Footer />
+        </body>
+      </NextIntlClientProvider>
     </html>
   );
 }