+++ src/components/uikit/Input/Input.tsx
@@ -1,8 +1,11 @@
-import { ComponentProps } from "react";
+import { ComponentPropsWithoutRef, ForwardedRef, forwardRef } from "react";
 import { input, type InputVariantProps } from "~/styled-system/recipes";
 
-type InputProps = InputVariantProps & ComponentProps<"input">;
+export type InputProps = InputVariantProps & ComponentPropsWithoutRef<"input">;
 
-export function Input({ variant, ...restProps }: InputProps) {
-  return <input className={input({ variant })} {...restProps} />;
-}
+export const Input = forwardRef(function Input(
+  { variant, ...restProps }: InputProps,
+  ref: ForwardedRef<HTMLInputElement>,
+) {
+  return <input ref={ref} className={input({ variant })} {...restProps} />;
+});