+++ src/components/uikit/index.ts
@@ -5,3 +5,6 @@ export * from "./Tag";
 export * from "./Avatar";
 export * from "./AvatarGroup";
 export * from "./Card";
+export * from "./Dropdown";
+export * from "./Select";
+export * from "./TextCollapse";