+++ src/pages/[locale]/events/create.tsx
@@ -26,7 +26,10 @@ export default function EventCreate() {
           property="og:description"
           content={t("metadata.events.create.description")}
         />
-        <meta property="og:image" content="https://kinguru.io/img/brand.png" />
+        <meta
+          property="og:image"
+          content="https://eventify.today/img/brand.png"
+        />
       </Head>
       <Navbar />
       <NewEventStepper />